requirements.txt
requests
speechrecognition
pyaudio
elevenlabs
langchain
langchain-community
python-dotenv
pydub
ffmpeg-python
